updated sharing
